+++ lib/weather_view.dart
@@ -1,0 +1,105 @@
+import 'package:flutter/material.dart';
+import 'package:flutter_svg/svg.dart';
+import 'package:yumemi_weather/yumemi_weather.dart';
+
+class WeatherView extends StatefulWidget {
+  const WeatherView({
+    super.key,
+  });
+
+  @override
+  State<WeatherView> createState() => _MainAppState();
+}
+
+class _MainAppState extends State<WeatherView> {
+  String _weather = '';
+
+  void getWeather() {
+    final yumemiWeather = YumemiWeather();
+    final weatherCondition = yumemiWeather.fetchSimpleWeather();
+    setState(() {
+      _weather = weatherCondition;
+    });
+  }
+
+  @override
+  Widget build(BuildContext context) {
+    return Scaffold(
+      body: Center(
+        child: FractionallySizedBox(
+          widthFactor: 0.5,
+          child: Column(
+            mainAxisAlignment: MainAxisAlignment.center,
+            children: [
+              Flexible(
+                child: Container(),
+              ),
+              Column(
+                mainAxisAlignment: MainAxisAlignment.center,
+                children: [
+                  AspectRatio(
+                    aspectRatio: 1,
+                    child: _weather != ''
+                        ? SvgPicture.asset('assets/$_weather.svg')
+                        : const Placeholder(),
+                  ),
+                  const Padding(
+                    padding: EdgeInsets.symmetric(vertical: 16),
+                    child: Row(
+                      mainAxisAlignment: MainAxisAlignment.spaceAround,
+                      children: [
+                        Text(
+                          '** ℃',
+                          textAlign: TextAlign.center,
+                          style: TextStyle(color: Colors.blue),
+                        ),
+                        Text(
+                          '** ℃',
+                          textAlign: TextAlign.center,
+                          style: TextStyle(color: Colors.red),
+                        ),
+                      ],
+                    ),
+                  ),
+                ],
+              ),
+              Flexible(
+                child: Column(
+                  children: [
+                    const SizedBox(height: 80),
+                    Padding(
+                      padding: const EdgeInsets.symmetric(vertical: 16),
+                      child: Row(
+                        mainAxisAlignment: MainAxisAlignment.spaceAround,
+                        children: [
+                          TextButton(
+                            onPressed: () {
+                              Navigator.pop(context);
+                            },
+                            child: const Text(
+                              'Close',
+                              textAlign: TextAlign.center,
+                              style: TextStyle(color: Colors.blue),
+                            ),
+                          ),
+                          TextButton(
+                            onPressed: getWeather,
+                            child: const Text(
+                              'Reload',
+                              textAlign: TextAlign.center,
+                              style: TextStyle(color: Colors.blue),
+                            ),
+                          ),
+                        ],
+                      ),
+                    ),
+                  ],
+                ),
+              ),
+            ],
+          ),
+        ),
+      ),
+    );
+  }
+}
